server/requirements.txt
amqp==2.5.2
asgiref==3.2.7
beautifulsoup4==4.9.0
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.6
django-celery-beat==2.0.0
django-cors-headers==3.3.0
django-redis==4.11.0
django-six==1.0.4
django-timezone-field==4.0
djangorestframework==3.11.0
fcm-django==0.3.4
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
kombu==4.6.8
lxml==4.5.0
psycopg2-binary==2.8.5
pyfcm==1.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.0
requests==2.23.0
six==1.15.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
xlwt==1.3.0
zipp==3.1.0
